bot/whatsappbot/requirements.txt
openai 
langchain
langchain_core
streamlit
langchain_community
langserve
langchain-openai
langchain_pinecone
pydantic
python-dotenv
pinecone-client 
pymongo
mysql-connector-python
langchainhub
sentence_transformers
langchain-groq
groq
fastapi
pydantic
fastapi-cors
uvicorn
twilio